requirements.txt
flask
azure-iot-device
